refactor session dirs
