+++ app/src/features/sessionBook/components/SessionPlayer/sessionPlayer.scss
@@ -17,7 +17,7 @@
     }
 
     .rr-controller {
-      height: auto;
+      height: auto !important;
 
       .rr-timeline {
         width: 100%;
