+++ app/src/features/sessionBook/components/SessionViewerBottomSheet/SessionViewerBottomSheet.tsx
