+++ app/src/features/sessionBook/components/SessionViewerBottomSheet/components/SessionInfo/SessionInfo.tsx
