+++ app/src/features/sessionBook/components/SessionViewerBottomSheet/components/SessionInfo/sessionInfo.scss
