+++ app/src/features/sessionBook/components/SessionsTitle/SessionTitle.tsx
