+++ app/src/features/sessionBook/components/SessionsTitle/sessionsTitle.scss
