+++ app/src/features/sessionBook/index.ts
@@ -1,3 +1,4 @@
-export { sessionRoutes } from "./routes";
 export * from "./types";
 export { SessionsOnboardingView } from "./screens/SessionsListScreen/components/OnboardingView/SessionsOnboardingView";
+export { SavedSessionScreen } from "./screens/SavedSessionScreen/SavedSessionScreen";
+export { sessionRoutes } from "./routes";
