+++ app/src/features/sessionBook/routes.tsx
@@ -1,9 +1,10 @@
 import { Navigate, RouteObject } from "react-router-dom";
 import PATHS from "config/constants/sub/paths";
-import { DraftSessionViewer, SavedSessionViewer } from "views/features/sessions/SessionViewer";
+import { DraftSessionViewer } from "views/features/sessions/SessionViewer";
 import NetworkSessionViewer from "views/features/sessions/SessionsIndexPageContainer/NetworkSessions/NetworkSessionViewer";
 import { SessionsListScreenContainer } from "./screens/SessionsListScreen/SessionsListScreenContainer";
 import SessionsFeatureContainer from "./container";
+import { SavedSessionScreen } from "features/sessionBook";
 
 export const sessionRoutes: RouteObject[] = [
   {
@@ -20,7 +21,7 @@ export const sessionRoutes: RouteObject[] = [
       },
       {
         path: PATHS.SESSIONS.SAVED.RELATIVE + "/:id",
-        element: <SavedSessionViewer />,
+        element: <SavedSessionScreen />,
       },
     ],
   },
