+++ app/src/features/sessionBook/screens/SavedSessionScreen/components/SavedSessionViewer/savedSessionViewer.scss
