--- app/src/features/sessionsV2/index.ts
@@ -1,1 +1,0 @@
-export { SavedSessionScreen } from "./screens/SavedSessionScreen/SavedSessionScreen";
