--- app/src/features/sessionsV2/types/index.ts
@@ -1,5 +1,0 @@
-export enum PlayerState {
-  PLAYING = "playing",
-  PAUSED = "paused",
-  SKIPPING = "skipping",
-}
