+++ app/src/src-SessionBear/routes/sessionRoutes.tsx
@@ -2,7 +2,7 @@ import { Navigate, RouteObject } from "react-router-dom";
 import PATHS from "config/constants/sub/paths";
 import { DraftSessionViewer } from "views/features/sessions/SessionViewer";
 import NetworkSessionViewer from "views/features/sessions/SessionsIndexPageContainer/NetworkSessions/NetworkSessionViewer";
-import { SavedSessionScreen } from "features/sessionsV2";
+import { SavedSessionScreen } from "features/sessionBook";
 import { SessionsListScreenContainer } from "features/sessionBook/screens/SessionsListScreen/SessionsListScreenContainer";
 import SessionsFeatureContainer from "features/sessionBook/container";
